feat: unify unix/windows self replace implementations
