+++ self-replacer/tests/integration_tests.rs
@@ -6,18 +6,13 @@
 use assert_cmd::Command;
 use predicates::prelude::*;
 use std::fs;
-use std::process::Command;
 use tempfile::TempDir;
 
 mod test_helpers;
+use self_replacer::{BinarySelfReplacer, SelfReplacer};
 use test_helpers::{copy_example_binary, create_modified_binary};
 
-#[cfg(unix)]
-use self_replacer::{SelfReplacer, UnixSelfReplacer};
-
 use self_replacer::BACKUP_SUFFIX;
-#[cfg(windows)]
-use self_replacer::{SelfReplacer, WindowsSelfReplacer};
 
 // ============================================================================
 // Common tests that run on all platforms
@@ -110,11 +105,7 @@ fn test_rollback_on_invalid_path() {
         test_dir.join("does_not_exist")
     };
 
-    #[cfg(unix)]
-    let result = UnixSelfReplacer::self_replace(&non_existent);
-
-    #[cfg(windows)]
-    let result = WindowsSelfReplacer::self_replace(&non_existent);
+    let result = BinarySelfReplacer::self_replace(&non_existent);
 
     assert!(result.is_err(), "Should fail when new binary doesn't exist");
